FT-475 order management api refactor code

+++ API/Controllers/ProductsController.cs
@@ -4,6 +4,7 @@ using Domain.Services;
 using Infrastructure.Models;
 using Microsoft.AspNetCore.Mvc;
 using System;
+using System.Collections.Generic;
 using System.Net;
 using System.Threading.Tasks;
 using Utilities.GlobalHelpers;
@@ -32,7 +33,7 @@ namespace API.Controllers
                         "Product can not be null")
                     );
             }
-
+            product.CreateDate = DateTime.Now;
             var result = await _productService.CreateProductAsync(product);
             var isSuccess = result.Item1;
             message = result.Item2;
@@ -41,7 +42,8 @@ namespace API.Controllers
             {
                 StatusCode = (int)HttpStatusCode.Created,
                 IsSuccess = true,
-                Message = $"{message}"
+                Message = $"{message}",
+                Data = product
             });
 
             return BadRequest(new
@@ -57,28 +59,61 @@ namespace API.Controllers
         {
             var products = await _productService.GetListProductsAsync();
 
-            return Ok(new
+            if (products != null)
             {
-                StatusCode = (int)HttpStatusCode.OK,
+                return Ok(new
+                {
+                    StatusCode = (int)HttpStatusCode.OK,
+                    IsSuccess = true,
+                    Message = "Get list success",
+                    Data = products,
+                });
+            }
+
+            return NotFound(new
+            {
+                StatusCode = (int)HttpStatusCode.NotFound,
                 IsSuccess = true,
-                Message = "Get list success",
-                Data = products,
+                Message = "Get list fail !",
             });
+
         }
 
         [HttpPut]
         public async Task<IActionResult> Update(ProductDto productDto)
         {
+            if (productDto == null)
+            {
+                return NotFound(new Error<string>(
+                        (int)HttpStatusCode.NotFound,
+                        "Create Failed",
+                        "Product can not be null")
+                    );
+            }          
+            var searchResult = await _productService.GetProductDtoByIdAsync(productDto.Id);
+            var productEntity = searchResult.Item1;
+            if (productEntity == null)
+            {
+                return NotFound(new
+                {
+                    StatusCode = (int)HttpStatusCode.NotFound,
+                    IsSuccess = true,
+                    Message = "Product that needs updating cannot found !",
+                });
+            }
+            productDto.ModifiedDate = DateTime.Now;
             var result = await _productService.UpdateProductAsync(productDto);
             var isSuccess = result.Item1;
             var message = result.Item2;
+            var data = new List<ProductDto> {productEntity, productDto };
             if (isSuccess)
             {
                 return Ok(new
                 {
                     StatusCode = (int)HttpStatusCode.OK,
                     IsSuccess = true,
-                    Message = $"{message}",                    
+                    Message = $"{message}", 
+                    Data = data
                 });
             }
             return BadRequest(new
@@ -101,7 +136,19 @@ namespace API.Controllers
                     IsSuccess = false,
                     Message = "Product Id Is InValid ! Delete Failed !"
                 });
-            }    
+            }
+
+            var tuple = _productService.GetProductDtoByIdAsync(new Guid(productId));
+            var productObject = tuple.Result.Item1;
+            if (productObject == null)
+            {
+                return NotFound(new
+                {
+                    StatusCode = (int)HttpStatusCode.NotFound,
+                    IsSuccess = false,
+                    Message = $"Product not found ! Delete Fail !",
+                });
+            }
             var result = await _productService.DeleteProductAsync(new Guid(productId));
             var isSuccess = result.Item1;
             var message = result.Item2;
@@ -112,6 +159,7 @@ namespace API.Controllers
                     StatusCode = (int)HttpStatusCode.NoContent,
                     IsSuccess = true,
                     Message = $"{message}",
+                    Data = productObject
                 });
             }
             return BadRequest(new
